Rename SmartObjectLayer API methods
