--- src/psd_tools/user_api/layers.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import (absolute_import, unicode_literals, print_function,
-    division)
