+++ tests/test_layer_masks.py
@@ -111,7 +111,7 @@ def test_mask_data_as_pil(filename):
         for l in layers:
             if l.has_mask():
                 mask = l.mask
-                if mask.is_valid:
+                if mask.is_valid():
                     assert mask.as_PIL() is not None
                 else:
                     bbox = mask.bbox
